requirements.txt
# Core Data Engineering
kafka-python==2.0.2
pyspark==3.5.0
pandas==2.1.3
numpy==1.26.2

# Database & Storage
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# Data Sources & APIs
requests==2.31.0
yfinance==0.2.32
alpha-vantage==2.3.1
python-binance==1.0.19

# Data Processing
pyarrow==14.0.1
fastparquet==2023.10.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Visualization
streamlit==1.29.0
plotly==5.18.0
dash==2.14.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
faker==20.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
schedule==1.2.0